Revert "refactor: Split routing utilities to allow testing without `getCollection`"
This reverts commit 8161114e3cf8951cf529df59048476dc7a1596e3.
